feat(materias): add credits field and fix creditos type in DB
- DB: changes creditos from integer to numeric(5,2) to support decimal values
- Model: adds credits: number | null to Subject interface
- Form: adds optional credits input (step 0.01) to add and inline edit forms
- Table: adds Créditos column; shows — when null

+++ database/db_structure.sql
@@ -52,7 +52,7 @@ create table materias
     id              integer generated always as identity primary key,
     nombre          varchar(120) not null,
     clave           varchar(20),
-    creditos        integer,
+    creditos        numeric(5, 2),
     cuatrimestre    integer,
     plan_estudio_id integer      not null,
     is_active       boolean                  default true,
